requirements.txt
matplotlib
numpy 
pandas
PyYAML
requests
scikit-learn 
seaborn 
streamlit 
xgboost